[router] initial scaffold
does the initial scaffolding for sanctuary-router, based on nixos-sbc.
we've got a bit of custom stuff in here until i figure out what i want
to do about the kernel -- upstream, per <github:nakato/nixos-sbc#34>,
does not want to use autoModules since it produces a large kernel that
takes a lot of resources to build (it's a ~1h build on a c4a-4core
machine on gcp for me).

The kernel configuration listed here seems to be misconfigured for boot,
cause it can't find the root disk (guessing i accidentally disabled
something).

For now, we'll keep this kernel config for future reference, and then
switch back to upstream's in the next commit or two, since at least they
enabled the right stuff for netfilter again.